visualization-pipeline/pyproject.toml
[project]
name = "zebra-visualization-pipeline"
version = "0.1.0"
description = "UMAP dimensionality reduction for embedding visualization"
requires-python = ">=3.11"
dependencies = [
    "umap-learn>=0.5.5",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
